Address comment: 1. Consolidate conformer stream step into mixin. 2. Consolidate streaming inference into mixin 3. Remove unused option in config.yaml 3. Consolidate strip_lang_tag into unified function. 4. Make lang_tag pattern customizable
