+++ nemo/collections/asr/parts/mixins/__init__.py
@@ -19,6 +19,7 @@ from nemo.collections.asr.parts.mixins.mixins import (
     ASRBPEMixin,
     ASRModuleMixin,
     DiarizationMixin,
+    PromptStreamingMixin,
 )
 from nemo.collections.asr.parts.mixins.multitalker_asr_mixins import SpeakerKernelMixin
 from nemo.collections.asr.parts.mixins.transcription import (
@@ -35,6 +36,7 @@ __all__ = [
     'ASRTranscriptionMixin',
     'DiarizationMixin',
     'InterCTCMixin',
+    'PromptStreamingMixin',
     'SpeakerKernelMixin',
     'TranscribeConfig',
     'TranscriptionMixin',
