+++ nemo/collections/asr/parts/mixins/mixins.py
@@ -591,7 +591,8 @@ class ASRModuleMixin(ASRAdapterModelMixin):
     
     def _apply_prompt_to_encoded(self, encoded: Tensor) -> Tensor:
         """Hook for prompt-conditioned subclasses to inject a language prompt
-        into the encoder output. Default: no-op."""
+        into the encoder output. Default: no-op. See ``PromptStreamingMixin``
+        for the prompt-aware override."""
         return encoded
 
     def conformer_stream_step(
@@ -904,3 +905,68 @@ class DiarizationMixin(VerificationMixin):
             Speaker labels
         """
         pass
+
+class PromptStreamingMixin:
+    """Adds language-ID prompt conditioning to a cache-aware streaming ASR model.
+
+    Overrides ``ASRModuleMixin._apply_prompt_to_encoded`` so that
+    ``conformer_stream_step`` injects a one-hot language prompt into the
+    encoder output. The host class is expected to set ``self.concat``,
+    ``self.num_prompts``, and ``self.prompt_kernel`` during its own
+    ``initialize_prompt_feature``.
+    """
+
+    def set_inference_prompt(self, target_lang: str):
+        """
+        Set the language prompt for streaming inference.
+
+        Call this before ``conformer_stream_step`` to condition decoding on
+        a specific language, following the same pattern as
+        ``change_decoding_strategy``.
+
+        Args:
+            target_lang: A key from the model's ``prompt_dictionary``
+                         (e.g. ``"en-US"``, ``"auto"``).
+        """
+        prompt_dict = self.cfg.model_defaults.get('prompt_dictionary', {})
+        if target_lang not in prompt_dict:
+            available = list(prompt_dict.keys())
+            raise ValueError(
+                f"Unknown target language '{target_lang}'. "
+                f"Available: {available[:20]}{'...' if len(available) > 20 else ''}"
+            )
+        self._inference_prompt_index = prompt_dict[target_lang]
+        logging.info(f"Inference prompt set to '{target_lang}' (index {self._inference_prompt_index})")
+
+    def _apply_prompt_to_encoded(self, encoded: Tensor) -> Tensor:
+        """
+        Inject the language-ID prompt into encoder output during streaming.
+
+        ``encoded`` arrives as (B, D, T) from the encoder cache-aware step.
+        Returns the same shape after prompt concatenation + projection.
+        """
+        if not self.concat or not hasattr(self, '_inference_prompt_index'):
+            return encoded
+
+        encoded = encoded.transpose(1, 2)  # (B, D, T) -> (B, T, D)
+
+        batch_size, time_steps, _ = encoded.shape
+        prompt = torch.zeros(
+            batch_size,
+            time_steps,
+            self.num_prompts,
+            dtype=encoded.dtype,
+            device=encoded.device,
+        )
+        idx = torch.full(
+            (batch_size,),
+            self._inference_prompt_index,
+            dtype=torch.long,
+            device=encoded.device,
+        )
+        prompt.scatter_(2, idx.view(batch_size, 1, 1).expand(-1, time_steps, -1), 1.0)
+
+        out_dtype = encoded.dtype
+        encoded = self.prompt_kernel(torch.cat([encoded, prompt], dim=-1)).to(out_dtype)
+        return encoded.transpose(1, 2)  # (B, T, D) -> (B, D, T)
+
